# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LONGEVITY API",
    "version": "0.1.0",
    "description": "Health-score MVP — backend API"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "displayName",
          "birthDate",
          "sex",
          "chronoAge"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "sex": {
            "type": "string",
            "enum": [
              "m",
              "f"
            ]
          },
          "chronoAge": {
            "type": "number"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "getMe",
        "tags": [
          "User"
        ],
        "summary": "Get current user",
        "responses": {
          "200": {
            "description": "User object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/organizations/join": {
      "post": {
        "operationId": "joinOrganization",
        "tags": [
          "Insurer"
        ],
        "summary": "Link the current b2c user to an organization via join code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "joinCode"
                ],
                "properties": {
                  "joinCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not a b2c user"
          },
          "404": {
            "description": "Invalid join code"
          }
        }
      }
    },
    "/organizations/leave": {
      "post": {
        "operationId": "leaveOrganization",
        "tags": [
          "Insurer"
        ],
        "summary": "Unlink the current b2c user from their organization",
        "responses": {
          "204": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not a b2c user"
          }
        }
      }
    },
...
    "/account": {
      "delete": {
        "operationId": "deleteAccount",
        "tags": [
          "Account"
        ],
        "summary": "Delete account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Wrong password"
          }
        }
      }
    },
    "/account/export": {
      "get": {
        "operationId": "exportAccount",
        "tags": [
          "Account"
        ],
        "summary": "DSGVO Art. 20 data export — user, sources, samples, scoreSnapshots, shareTokens as JSON",
        "responses": {
          "200": {
            "description": "Full account data export (application/json, Content-Disposition: attachment)"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/account/consent": {
      "get": {
        "operationId": "getAccountConsent",
        "tags": [
          "Account"
        ],
        "summary": "Get GDPR Art. 9 health data consent status and text",
        "responses": {
          "200": {
            "description": "Consent status"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      },
      "post": {
        "operationId": "grantAccountConsent",
        "tags": [
          "Account"
        ],
        "summary": "Grant GDPR Art. 9 consent to process health data",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consent recorded"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/account/consent/revoke": {
      "post": {
        "operationId": "revokeAccountConsent",
        "tags": [
          "Account"
        ],
        "summary": "Revoke GDPR Art. 9 health data consent (Art. 7(3) GDPR)",
        "responses": {
          "200": {
            "description": "Consent revoked"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
...
  }
}